+++ frontend/src/pages/MainPage.vue
@@ -26,7 +26,6 @@
 
 /* eslint-disable import/no-unresolved */
 import airbnbs from "@/data/airbnb.json";
-import properties from "@/data/properties.json";
 import Heatmap from "@/components/map/Heatmap.vue";
 import PropertyCard from "@/components/PropertyCard.vue";
 
@@ -43,7 +42,6 @@ export default {
     return {
       center: { lat: 40.730610, lng: -73.935242 },
       airbnbs: airbnbs, // TODO get from backend
-      properties: properties,
       selected_property: null, // The property that was clicked
     };
   },
@@ -71,57 +69,9 @@ export default {
 
       return result;
     },
-
-    /**
-     * Geographical property distribution
-     * @returns {Array}
-     */
-    property_locations() {
-      const result = [];
-      this.properties.forEach((property) => {
-        result.push({
-          lat: property.latitude,
-          lng: property.longitude,
-          id: property.id,
-        });
-      });
-
-      return result;
-    },
-  },
-
-  created() {
-    // Get Google maps API
-    this.getMapsApi();
-
-    // Use google geocoding API to get location for all properties
-    this.properties.slice(0, 10).forEach((property) => {
-      // Determine full address string
-      const address_string = `${property.ADDRESS}, ${property["ZIP CODE"]}, New York`;
-
-      console.log("Query geocoding for property with address", property, address_string);
-
-      // Make query to geocoding API
-    });
   },
 
   methods: {
-    /**
-     * TODO
-     */
-    async getMapsApi() {
-      // eslint-disable-next-line global-require
-      const loadGoogleMapsApi = require("load-google-maps-api");
-      const apiKey = "***REMOVED***";
-      const options = {
-        key: apiKey,
-        libraries: ["visualization"],
-      };
-
-      const maps_api = await loadGoogleMapsApi(options);
-      console.log("Setting API!");
-      this.$store.commit("setMapsApi", {maps_api: maps_api});
-    },
 
     /**
      * TODO
